requirements.txt
absl-py==0.13.0
aiohttp==3.6.3
astunparse==1.6.3
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.10.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.0.5
clang==5.0
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
dblpy==0.4.0
discord-buttons-plugin==0.1.2
discord.py==1.7.3
dnspython==2.1.0
flatbuffers==1.12
fuzzywuzzy==0.18.0
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.40.0
h5py==3.1.0
idna==2.10
keras==2.6.0
Keras-Preprocessing==1.1.2
Markdown==3.3.4
multidict==4.7.6
numpy==1.19.5
oauthlib==3.1.1
opencv-contrib-python==4.5.1.48
opencv-python==4.5.1.48
opt-einsum==3.3.0
pandas==1.3.3
protobuf==3.18.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pymongo==3.11.3
PyNaCl==1.4.0
python-dateutil==2.8.2
python-Levenshtein==0.12.2
pytz==2021.1
ratelimiter==1.2.0.post0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
selenium==3.141.0
six==1.15.0
soupsieve==2.2.1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.6.0
tensorflow-gpu==2.6.0
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.6
webdriver-manager==3.4.2
Werkzeug==2.0.1
wrapt==1.12.1
yarl==1.5.1
youtube-dl==2021.6.6
